ux-analyzer-backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
networkx==3.5
numpy==2.2.6
packaging==25.0
pillow==11.3.0
scipy==1.16.1
SQLAlchemy==2.0.41
tifffile==2025.6.11
typing_extensions==4.14.0
Werkzeug==3.1.3
